% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bannerMerger.R
\name{bannerMerge}
\alias{bannerMerge}
\title{union of picture,especially adbanner}
\usage{
bannerMerge(df, textColor = "#FFFFFF", textBoxColor = "#00000070",
  imageIdentifier_textColor = "#FFFFFF",
  imageIdentifier_textBoxColor = "#000000200",
  imageIdentifier_textSize = 30, captionType = "outside",
  captionOutside_backgroundColor = "#cccccc", captionOutside_rectSize = 100,
  caption_textSize = 30)
}
\arguments{
\item{df}{dataframe contains image file path}

\item{textColor}{assignment of picture file name and picture width,height}

\item{textBoxColor}{assignment of backgournd color with textColor(param)}

\item{imageIdentifier_textColor}{assignment of color of picture identifier}

\item{imageIdentifier_textBoxColor}{assignment of backgroud color with imageIdentifier_textColor(param)}

\item{imageIdentifier_textSize}{assignment of text size of picture identifier}

\item{captionType}{assignment of image caption type,"outside" or "inside".in most case, "inside" is better.}

\item{captionOutside_backgroundColor}{when the captionType is "outside",assignment of backgroud color of rect}

\item{captionOutside_rectSize}{when the captionType is "outside",assignment of size of rect}

\item{caption_textSize}{when the captionType is "outside",assignment of text size of rect}
}
\description{
union of picture,especially adbanner
}
\author{
Kazuyoshi Nakakita
}
